requirements.txt
alembic==0.9.8
aniso8601==2.0.1
asn1crypto==0.24.0
attrs==17.4.0
cffi==1.11.5
click==6.7
coverage==4.5.1
cryptography==2.2.2
flake8==3.5.0
Flask==0.12.2
Flask-JWT-Extended==3.7.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.5.3
pytest==3.4.1
pytest-cov==2.5.1
pytest-cover==3.0.0
pytest-coverage==0.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.3
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.4
Werkzeug==0.14.1
